Maybe.php: Added getOrElseFn method
This method accepts a function that is called if the Maybe instance is
Nothing. The function should return a value. The benefit of this method
is that the function is only evaluated when its return value is needed.
It therefore is a better alternative for getJustOrElse if the fallback
value is expensive to calculate.